actions/answers tests : moved shared to helpers/shared
