+++ packages/@coorpacademy-app-player/src/actions/ui/test/answers-validation/helpers/shared.js
@@ -1,7 +1,6 @@
-import test from 'ava';
 import set from 'lodash/fp/set';
-import {UI_TOGGLE_ACCORDION} from '../../../ui/corrections';
-import {ANSWER_FETCH_REQUEST, ANSWER_FETCH_SUCCESS} from '../../../api/answers';
+import {UI_TOGGLE_ACCORDION} from '../../../../ui/corrections';
+import {ANSWER_FETCH_REQUEST, ANSWER_FETCH_SUCCESS} from '../../../../api/answers';
 
 export const accordionIsOpenAt = i => [
   [
@@ -37,7 +36,3 @@ export const fetchCorrection = [
     set('ui.answers.0.correction', 'Bonne réponse', {})
   ]
 ];
-
-test('passing tools', t => {
-  t.pass();
-});
